data_science/pyproject.toml
[project]
name = "data_science"
version = "0.1.0"
description = "Data platform demonstration that integrates dagster, dbt, sling, and dlt."
requires-python = ">=3.10, <3.11"
dependencies = [
    "dagster-cloud==1.11.12",
    "dagster-k8s==0.27.12",
    "dagster-snowflake==0.27.12",
    "faker==37.11.0",
    "llvmlite==0.41.1",
    "numba==0.58.1",
    "psutil==7.1.0",
    "snowflake==1.8.0",
    "snowflake-ml-python==1.11.0",
    "snowflake-snowpark-python==1.36.0",
    "xgboost-cpu==3.0.5"
]

[tool.uv]
override-dependencies = [
    "xgboost; sys_platform == 'never'"
]

[dependency-groups]
dev = [
    "dagster-dg-cli",
    "dagster-webserver",
    "jupyter",
    "mkdocs",
    "mkdocs-gen-files",
    "mkdocs-material",
    "mkdocs-mermaid2-plugin",
    "mkdocstrings",
    "mkdocstrings-python",
    "sphinx",
    "sqlfluff",
    "seaborn",
    "itables",
    "faker",
    "pandas",
    "tuna",
    "snowpark-extensions"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
mypkg = "data_science"

[tool.ruff]
line-length = 88

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 20
quote-style = "double"
indent-style = "space"
[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B",  # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
]

[tool.sqlfluff]
core.exclude_rules = ["AL07", "RF02", "ST06"]
core.max_line_length = 88
core.dialect = "snowflake"
indentation.tab_space_size = 4
layout.type.alias_expression.spacing_before = "align"
layout.type.alias_expression.align_within = "select_clause"
rules.aliasing.table.aliasing = "implicit"
rules.aliasing.column.aliasing = "implicit"

# # core.templater = "dbt"
# # templater.dbt.project_dir = "./dbt"
# # templater.dbt.profiles_dir = "./dbt"
# # experimental.format.executeInTerminal = true

[tool.dagster]
module_name = "data_science.definitions"
code_location_name = "data_science"

[tool.dg]
directory_type = "project"
project.root_module = "data_science"

